requirements.txt
pymupdf>=1.24.0
rank-bm25>=0.2.2
numpy>=1.26.0
requests>=2.31.0
aiohttp>=3.9.0
google-genai>=0.1.2
python-dotenv
